requirements.txt
mysql-connector-python==8.0.33
smartapi-python==1.1.0
pandas==2.0.3
python-dateutil==2.8.2
logzero==1.7.0
websocket-client==1.7.0